Mediator selection function

This function selects the mediators while controlling for FDR.

Usage

MediationFDR(
  X,
  Y,
  M,
  V1 = NULL,
  V2 = NULL,
  q1 = 0.025,
  q2 = 0.025,
  f_func = "Diff",
  correction_method = "holm",
  weighted = FALSE,
  binaryOutcome = FALSE
)

Arguments

X

exposure of interest

Y

outcome of interest

M

potential mediators

V1

confounding variables in X-M path

V2

confounding variables in M-Y path

q1

FWER level in step 1

q2

FDR controlled level in step 4

f_func

function form of W in step 4

correction_method

multiple comparison correction method in step 1

weighted

whether we weight the Z statistics in step 3

binaryOutcome

whether the outcome is binary

Value

selected mediators

Examples

set.seed(20205)
data = datagen(n = 1000, p = 100, a = 0.5, b = 0.3, rho = 0.1)
MediationFDR(X = data$X, Y = data$Y, M = data$M, V1 = NULL, V2 = NULL, q1 = 0.025, q2 = 0.025, f_func = "Diff", correction_method = "holm", weighted = FALSE)
#> $med_select
#>  [1]  6  7  8  9 10 11 12 13 14 15
#> 
#> $path.a
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15
#> 
#> $path.b
#>  [1]  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 22 38 42 86
#> 
#> $path.b.tilde
#> [1]  1 24 49 60 64 69 82 84
#> 
#> $path.b.weighta
#>  [1]  6  7  8  9 10 11 12 13 14 15
#> 
#> $path.b.tilde.weighta
#> [1] 1
#>