Data generation function

This function generates the simulation data (exposure X, outcome Y, and potential mediators M).

For each subject i = 1, ..., n:

\(X_i \stackrel{i.i.d}{\sim} Bernoulli(prob = 0.5)\)

\(M_{i,k} = \alpha_k X_i + e_{1_{i,k}}\), where \(e_{1_{i,k}} \stackrel{i.i.d}{\sim} N(0, \Sigma), k = 1, ..., p, \Sigma\) is the variance-covariance matrix with diagonal entries 1 and off-diagonal entries rho

\(Y_i = X_i + \sum_k b_{k} M_{i,k} + e_{2_{i}}\), where \(e_{2_{i}} \stackrel{i.i.d}{\sim} N(0, 1)\)

The 6th-15th M variables (M6,M7,M8,M9,M10,M11,M12,M13,M14,M15) are set to be the true mediators (i.e., having non-zero alpha and beta coefficients).

Specifically, the first 15 coefficients of alpha are set to take some non-zero value a, the 6th to 20th coefficients of beta are set to take some non-zero value b, and all other alpha and beta coefficients are set to be 0.

Usage

datagen(n, p, a, b, rho)

Arguments

n

sample size

p

number of potential mediators

a

effect size for X-M path (i.e., non-zero alpha coefficient)

b

effect size for M-Y path (i.e., non-zero beta coefficient)

rho

effect size for the correlation structure of M

Value

The resulting data has 3 lists: X, Y and M

Examples

set.seed(20205); data = datagen(n = 1000, p = 100, a = 0.5, b = 0.3, rho = 0.1)
head(data$X)
#> [1] 0 1 0 1 1 0
head(data$Y)
#>            [,1]
#> [1,]  0.1358174
#> [2,] -0.7465296
#> [3,]  0.9623505
#> [4,] -0.8921536
#> [5,]  4.0902124
#> [6,]  1.5694765
head(data$M)
#>              M1          M2          M3          M4         M5          M6
#> [1,]  0.5027504 -0.16291036  2.28875009 -0.22211516  1.4444724 -0.09632082
#> [2,]  2.8842757  0.67104604 -0.05653073 -0.03181477 -1.4822663  0.94140738
#> [3,]  0.8806485  2.52424586 -0.70099327 -0.19855981 -1.4190167 -0.07663935
#> [4,]  1.7024567 -0.37017428  0.58538969  1.01723077 -0.3836153  0.03433381
#> [5,]  1.3613710 -0.02651447 -0.25478171  2.56942075  0.9977460 -0.65373616
#> [6,] -0.5367336 -1.09813514 -0.41355160  0.48842438  1.1974492 -0.10119078
#>              M7        M8        M9         M10        M11        M12
#> [1,]  0.7671144 -0.804897 0.3262766 -0.04096526  0.1119893 -1.4802215
#> [2,]  0.5708139 -1.350358 1.2661176 -0.29800450  1.5093193 -1.7855316
#> [3,] -1.1650013  1.977935 1.3431788 -0.12970635 -0.9834703 -0.3147970
#> [4,] -0.7775283 -0.746832 0.9444932 -2.91674074  2.2641524  0.4704194
#> [5,]  1.9539779  1.764747 0.1429410  2.10091628  0.6983898  0.1990981
#> [6,] -0.9326512  0.632527 0.4831236  0.13215381 -1.0693534 -1.2281335
#>             M13        M14        M15        M16         M17        M18
#> [1,]  0.9515425  1.7344801  0.7283838  1.5096686 -0.70678343 -0.7804054
#> [2,]  0.4337681 -1.6928977  0.3748492 -1.0462682  0.48978747 -0.5498706
#> [3,]  1.6139587  1.2109272 -0.1288446  0.5108508 -0.12998839 -0.9022660
#> [4,] -0.3421771  0.6286474  0.8671723 -2.9163356 -1.56500872 -0.5651489
#> [5,]  0.8656623  2.5523550  0.1435542 -2.1236006  0.09157872  1.2678717
#> [6,] -0.6549213 -0.5925755 -2.5477310  0.7500170 -0.51299443  1.1257781
#>              M19         M20         M21         M22        M23         M24
#> [1,]  0.12823659 -0.40746886 -0.49327375  0.46140216  0.1733370 -1.44575321
#> [2,] -0.07716728  0.33338100  0.07615032 -2.21130520  0.8075281 -0.07639087
#> [3,]  1.38120342 -1.01202403  0.10542017 -0.05417739 -0.4255744 -0.52562429
#> [4,]  0.85898956 -0.86596349  0.06764030 -1.84735040 -0.2070313 -1.01033975
#> [5,] -0.30695813 -1.28854668 -0.36600262  0.65950367  0.7041140  1.63061670
#> [6,]  2.98663955  0.07652582  2.57231714  0.23321110  1.3348329  1.05416176
#>             M25        M26        M27         M28        M29        M30
#> [1,]  0.8688064  0.6605519  1.7138658  0.14957049  0.8805247  1.8291074
#> [2,]  0.6549450 -0.2526353 -0.5923664 -0.73525484  2.2310318  0.8751341
#> [3,]  1.0737954 -0.9129543 -0.9474339 -0.15020652  1.2811060  0.7752486
#> [4,] -2.1671731 -1.2010788  1.4954428 -0.07245362 -1.0927275 -0.5051496
#> [5,] -0.7421040 -0.8142610  1.3545506  0.97647643  0.1308949 -0.1582398
#> [6,] -0.4292953 -0.7008030  1.7662054 -0.23916112 -0.1777651 -1.7239776
#>              M31        M32        M33        M34         M35       M36
#> [1,]  0.01140657  0.2914104  0.5075838  0.3649896  1.08934915 0.5725834
#> [2,]  0.29888797  0.3297477 -1.7638165 -1.9386282  1.02824102 0.9925051
#> [3,] -0.29543844 -0.1356256  0.3631291  1.5739610 -0.88499943 0.4285352
#> [4,]  0.94666142 -1.0219818 -0.7460691  0.1643160 -0.90018543 0.7913280
#> [5,] -1.25717405 -0.3698548 -1.3538105 -0.6049487  0.27441794 1.1551193
#> [6,] -0.40708169  1.0731787  1.2274653  1.2187296 -0.04722578 1.9127676
#>             M37        M38        M39       M40        M41        M42
#> [1,] -1.5816528  1.4868981 -1.2102815 1.4202218 -0.8488640 -0.4892918
#> [2,] -0.4655932  0.4693938  1.0968342 0.5698962 -1.0671102 -0.7690780
#> [3,] -0.7758478 -1.0636917  0.9941119 0.9576331  1.5804221  0.4989825
#> [4,]  0.3452347 -0.4168524 -1.6969934 0.6369699 -0.6090653  0.5148267
#> [5,] -0.1302863 -0.8653963  0.4983362 2.0693718  2.2487388  0.7790776
#> [6,] -1.1491320 -1.2413385  1.0659883 1.6874090  0.6976956  2.1455033
#>             M43         M44        M45         M46        M47        M48
#> [1,]  1.3387938  0.92453777  1.7560679 -1.82493613 -0.5738911 -0.4979661
#> [2,] -1.5209640  0.97001854  0.5107750  0.16980737  0.5629689  1.1825584
#> [3,] -0.2222986  0.41851353 -1.2354896  0.28520678 -0.1803093  1.2498444
#> [4,] -0.8328782 -1.08033563 -0.5954615  0.05560754  0.8396333  0.1940971
#> [5,]  0.1640344  0.05609372 -0.2017638 -0.12226275  0.0823320 -0.1687620
#> [6,]  0.6146529  0.72977946  1.0094568  0.79106197  1.1246688 -0.1914467
#>              M49        M50        M51        M52         M53        M54
#> [1,]  1.05664644  0.4525241  0.1930294  1.1768248 -1.39369903  0.2672023
#> [2,]  0.84221869 -0.7548134  1.4775800 -0.8570372  0.07606974 -1.6469116
#> [3,]  1.14703108 -0.2019577 -1.1448762  1.8974470 -0.29395616 -0.4386511
#> [4,] -0.45951259  0.4075782  0.3123060  0.6826160 -0.75796366  0.9048632
#> [5,]  0.17670981  2.3286003  0.3688321 -0.8020731 -0.79867582  0.8834646
#> [6,] -0.03988408 -1.4196069  0.5271771 -1.1263211  0.42871027  0.4461303
#>             M55        M56          M57        M58        M59        M60
#> [1,]  0.1051527  1.3538501  0.628765970 -1.4315778 -1.0080418  1.6323990
#> [2,] -0.6637260 -0.1114877 -1.224397472  0.1827549 -0.5149444 -0.5323562
#> [3,] -0.6043152 -0.4994222  0.779507524 -0.1623187 -0.3549908  2.1106864
#> [4,] -0.5498873 -2.1514475 -1.799914407 -2.1525857 -0.5718581 -0.5408472
#> [5,] -0.8078095  1.8013760  0.007164344 -1.8994773  0.2331146 -1.0067394
#> [6,]  1.7182077 -0.5063559  1.778488544  1.2099194  0.1874147 -1.1830699
#>              M61        M62        M63        M64         M65        M66
#> [1,]  1.29784039  0.5083949 -0.1727160 -1.0070135  0.07864695 -1.2413463
#> [2,]  0.58110911 -1.2494192 -1.4156795  1.2283734  0.24846525  0.5505654
#> [3,]  0.02975303 -0.1631824 -2.4030992  0.7146001  1.05792359 -0.6686169
#> [4,] -0.60313470 -0.2529380 -0.6931866 -0.3864072 -0.18675030  0.6161867
#> [5,] -0.55415737  1.4666995 -0.6539787  0.5514469  1.51749306  1.7960789
#> [6,]  0.47727611 -0.2627578  2.0948647  0.1170738  1.51410618  1.8891771
#>             M67        M68        M69        M70        M71         M72
#> [1,] -0.2082777  0.1953478  2.2462927  1.1239853 -0.8886593 -0.49535167
#> [2,] -0.4130883  0.1691805 -0.9707836  0.4673199 -0.2183907 -1.41996390
#> [3,]  1.6805910  1.1422207 -1.9442055 -1.1183200  0.7692355  0.12626507
#> [4,] -1.6502159 -0.1510277 -0.6981349  0.6134938  0.3721059  0.06944933
#> [5,]  1.2793783 -0.5184597  0.4482489  0.9949862 -0.5970378  0.48951290
#> [6,]  0.7242190  0.1036103  0.3457743  1.2104396  1.1233600 -0.23605563
#>              M73        M74        M75        M76          M77         M78
#> [1,]  1.34116180 -0.3714890  1.1861243  0.7943543 -0.233968021  0.89441013
#> [2,] -1.60037258 -1.3360349 -0.7946631  1.2736373 -0.933725246 -0.62221499
#> [3,] -0.88011148  0.6433123  0.6015969  2.7644875 -0.897178078 -0.30367469
#> [4,] -0.04072279 -0.1655819 -0.3292523  0.1723502 -0.009659608 -1.22131385
#> [5,]  0.45858911  0.6302948  0.1602041 -0.8151253  0.073017945  0.60094135
#> [6,]  0.58738767  0.4359742 -1.5765565  2.1449710  0.469621974  0.04093697
#>             M79        M80         M81          M82        M83        M84
#> [1,] -0.1907340 -0.4515134  0.01089041  0.532557612  0.4515331  0.2129983
#> [2,] -1.2006326 -0.8535195 -0.60614933  0.200645383  0.4768520 -0.5241239
#> [3,] -0.6121369  1.3761576  0.69613107  0.757212036 -0.9932301  0.6927489
#> [4,] -0.2285029 -0.6845087 -0.09296953 -0.002010584 -0.1240725 -1.3831319
#> [5,]  1.2039117 -0.9619433  0.82208498 -0.123498317  0.2613388  1.7710743
#> [6,]  0.9982752 -2.0204699 -0.38247679 -0.508292466 -0.8101829 -0.6172028
#>             M85        M86        M87        M88        M89        M90
#> [1,]  1.4048001  1.8275532 -1.2579331 -2.3508205 -0.2853930  0.3839088
#> [2,] -0.8220202  0.2277155 -0.7175042 -1.3510001  0.5126634  0.4680439
#> [3,] -0.4062063 -0.7159284 -0.1743167 -0.4224354 -1.1480502 -0.8299133
#> [4,] -0.5445979 -2.8517571 -0.1393255 -0.3137473  0.2859264  0.6778276
#> [5,]  1.8688637  0.1041730 -0.2828779  0.3604942  0.3101008  0.1471100
#> [6,]  1.2153132  0.1296473 -0.5499495 -1.2385831  1.3352138  0.3650290
#>             M91         M92        M93         M94        M95         M96
#> [1,]  0.5000214 -0.05595704  1.2806533  1.20448739  0.8926552 -0.09007026
#> [2,] -1.6928767 -1.74691315 -0.2434929 -0.21739272 -0.8100143 -0.58806234
#> [3,]  1.7341601  1.19262129  0.6100946 -0.04011231 -1.5495743  1.68171414
#> [4,] -1.7111338 -0.56350133 -1.0656337 -2.37834434 -0.2284628  1.74243962
#> [5,]  0.2144922  0.85292880 -0.1562005  0.57388089 -1.1696349 -0.26759877
#> [6,]  1.4726883 -0.92639694  0.6045032  0.71380803  1.5039876  0.34378099
#>              M97         M98         M99       M100
#> [1,]  0.02330135 -0.42370277 -0.04844130  0.6134511
#> [2,] -0.27441857 -0.98498103  0.66076233 -1.2673000
#> [3,]  0.26543242 -0.90705952  0.57590889  1.4259155
#> [4,] -0.01262850 -0.36885479 -0.42566648  0.1728041
#> [5,] -0.24999333 -0.04904358  1.03188373  0.9297149
#> [6,]  0.08299186  0.52576547  0.02660241  1.2683944