Data generation function for counter example

This function generates the simulation data (exposure X, outcome Y, potential mediators M, and covariate V1) for counter example.

The number of potential mediators is set to be p = 19.

For each subject i = 1, ..., n:

\(X_i \stackrel{i.i.d}{\sim} N(0, 1)\)

\({V_1}_i \stackrel{i.i.d}{\sim} N(0, 1)\)

\(cov(X, V_1) = 0.8\)

\(M_{i,k} = \alpha_k X_i + {V_1}_i {\eta_1}_k + e_{1_{i,k}}\), where \(e_{1_{i,k}} \stackrel{i.i.d}{\sim} N(0, 1), k = 1, ..., p\), and \(cov(e_{1_{i,k}},e_{1_{i,j}} = 0)\) for k not equal j

\(Y_i = X_i + \sum_k \beta_{k} M_{i,k} + e_{2_{i}}\), where \(e_{2_{i}} \stackrel{i.i.d}{\sim} N(0, 1)\)

We let \(\alpha_1, \cdots, \alpha_{10}\) be randomly from the uniform distribution from -1 to 1 and \(\alpha_{11}, \cdots, \alpha_{19}\) be 0.

We let \({\eta_1}_{11} = 1\) and the rest of the \(\eta_1\) values be 0.

We let \(\beta_1, \cdots, \beta{9} = 0\) and \(\beta_{10}, \cdots, \beta{19}\) be either 1 or -1, randomly.

M10 is the true mediators (i.e., having non-zero alpha and beta coefficients).

Usage

datagen.counter(n)

Arguments

n

sample size

Value

The resulting data has 4 lists: X, Y, M, V1

Examples

set.seed(23)
data = datagen.counter(n = 1000)
head(data$X)
#>             [,1]
#> [1,] -0.02158138
#> [2,]  1.61887846
#> [3,]  1.38667551
#> [4,]  0.20707462
#> [5,]  0.74535031
#> [6,] -0.27278171
head(data$Y)
#>            [,1]
#> [1,]  1.0435139
#> [2,] -0.2097866
#> [3,]  1.9088426
#> [4,]  4.1112320
#> [5,]  0.4821536
#> [6,]  3.3867542
head(data$M)
#>              M1         M2         M3          M4         M5         M6
#> [1,] -1.2256408  1.1419768  1.3049401 -1.45855173  0.3148073 -0.8549353
#> [2,]  0.4658685  0.7749704 -0.4386567 -0.21061086  1.6264951  0.1302713
#> [3,] -0.1527225  0.5862869  0.2699257  0.22264221  0.3244481 -0.6264304
#> [4,] -2.0035015 -1.3615076  0.3444736 -0.07934474 -0.2234448 -0.9655403
#> [5,]  0.8838462 -0.3606156 -2.2570514  1.18961311 -0.6892579 -2.1474007
#> [6,]  0.3454013  0.1318351  0.2379899  0.15684527 -0.8283692 -0.4469036
#>               M7          M8         M9         M10       M11        M12
#> [1,]  1.14250457 -1.54428095  1.1061722 -0.52240556 0.6627643  1.2400793
#> [2,]  1.67418614  0.15069038 -0.5454678  0.02726723 2.0226588 -0.8044814
#> [3,] -0.04889471 -0.04207574 -0.2651397 -0.19133908 1.7051222  0.2159686
#> [4,] -0.64222672 -0.06010606 -0.7196775  0.41883397 0.2807591 -0.7554807
#> [5,] -1.12913004 -1.08489811 -1.0501237  0.02716947 2.4495890 -0.8875516
#> [6,] -0.17651224  0.88816886  1.7812724  0.98816977 0.1787698 -0.9811137
#>             M13         M14         M15         M16        M17        M18
#> [1,]  0.2742722  0.29376461 -1.06084797 -0.86693817  0.6225121  1.7529229
#> [2,] -0.8803008 -0.54220915  0.34068504  0.09062461 -1.0396295 -0.7315671
#> [3,] -0.1246070 -0.15405034  0.03452526  1.00886900 -0.5039592 -1.4885764
#> [4,] -0.2380876  1.32211001 -2.25948456 -1.15507455 -0.2790671 -0.5794965
#> [5,]  0.3692004 -0.01443777 -2.03971243 -1.60878163  0.8863455  0.2435306
#> [6,] -0.5205271  0.88760429 -0.06916727 -0.95773789  1.2067612  0.6582457
#>             M19
#> [1,]  0.3865076
#> [2,] -1.0041114
#> [3,] -1.2114651
#> [4,] -0.4420720
#> [5,]  0.5061796
#> [6,] -0.4733559
head(data$V1)
#>            [,1]
#> [1,] -0.3023843
#> [2,]  2.0124805
#> [3,]  1.4362650
#> [4,]  0.7872411
#> [5,]  1.4297402
#> [6,] -0.8384281